flowing fix, pasteur prog fix, presets edited

+++ include/Core/TimeObj.hpp
@@ -13,22 +13,22 @@ class TimeObj
 {
 protected:
     STM32RTC& rtc = STM32RTC::getInstance();
-    uint8_t hours = 0;
-    uint8_t minutes = 0;
-    uint8_t seconds = 0;
-    uint8_t days = 0;
-    uint8_t months = 0;
-    uint8_t years = 0;
+    uint64_t hours = 0;
+    uint64_t minutes = 0;
+    uint64_t seconds = 0;
+    uint64_t days = 0;
+    uint64_t months = 0;
+    uint64_t years = 0;
 
 public:
-    uint8_t getSecs();
-    uint8_t getMins();
-    uint8_t getHours();
-    uint8_t getDays();
-    uint8_t getMonths();
-    uint8_t getYears();
+    uint64_t getSecs();
+    uint64_t getMins();
+    uint64_t getHours();
+    uint64_t getDays();
+    uint64_t getMonths();
+    uint64_t getYears();
     void clone(TimeObj *parent, CloneType type = CloneType::All);
-    void addMinutes(uint8_t minutes);
+    void addMinutes(uint64_t minutes);
     bool isAnotherDay(TimeObj *reference);
     bool isBiggerThan(TimeObj *reference, bool relative_today = false);
     bool isTimeEqual(TimeObj *reference);
